+++ crates/client/db/src/error.rs
@@ -1,44 +1,26 @@
-use std::{error, fmt};
-
 use bonsai_trie::DBError;
+use thiserror::Error;
+
+use crate::Column;
 
 #[derive(thiserror::Error, Debug)]
 pub enum DbError {
     #[error("Failed to commit DB Update: `{0}`")]
-    CommitError(#[from] sp_database::error::DatabaseError),
+    RocksDB(#[from] rocksdb::Error),
     #[error("Failed to deserialize DB Data: `{0}`")]
     DeserializeError(#[from] parity_scale_codec::Error),
     #[error("Failed to build Uuid: `{0}`")]
     Uuid(#[from] uuid::Error),
     #[error("A value was queryied that was not initialized at column: `{0}` key: `{1}`")]
-    ValueNotInitialized(u32, String),
+    ValueNotInitialized(Column, String),
 }
 
-#[derive(Debug)]
+#[derive(Debug, Error)]
 pub enum BonsaiDbError {
-    Io(std::io::Error),
-}
-
-impl fmt::Display for BonsaiDbError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            BonsaiDbError::Io(ref err) => write!(f, "IO error: {}", err),
-        }
-    }
-}
-
-impl error::Error for BonsaiDbError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        match *self {
-            BonsaiDbError::Io(ref err) => Some(err),
-        }
-    }
-}
-
-impl From<std::io::Error> for BonsaiDbError {
-    fn from(err: std::io::Error) -> BonsaiDbError {
-        BonsaiDbError::Io(err)
-    }
+    #[error("IO error: `{0}`")]
+    Io(#[from] std::io::Error),
+    #[error("Failed to commit DB Update: `{0}`")]
+    RocksDB(#[from] rocksdb::Error),
 }
 
 impl DBError for BonsaiDbError {}