+++ crates/node/src/service.rs
@@ -180,7 +180,7 @@ where
         telemetry.as_ref().map(|x| x.handle()),
     )?;
 
-    let deoxys_backend = DeoxysBackend::open(&config.database, &db_config_dir(config), cache_more_things)?;
+    let deoxys_backend = DeoxysBackend::open(&config.database, &db_config_dir(config), cache_more_things).unwrap();
 
     let (import_queue, block_import) = build_import_queue(
         client.clone(),